% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bioclimate}
\alias{bioclimate}
\title{Topoclimate estimates}
\usage{
bioclimate(dem, include_inputs = FALSE)
}
\arguments{
\item{dem}{RasterLayer representing elevation, in meters. A spatial resolution of 10-30 m is recommended. The dataset must be in the continental US or Canada. DEMs covering areas larger than a small landscape may encounter computational challenges.}

\item{include_inputs}{Whether to return intermediate inputs in addition to topoclimate result (default FALSE)}
}
\value{
A raster stack of topoclimate variables, including "high_temperature" (in deg C), "low_temperature" (in deg C), and "moisture" (in mm).
}
\description{
This function produces estimates of biologically effective microclimate given a user-suppled digital elevation model (DEM). It is based on a model that leverages North American tree species occurrences as microclimate indicators. Estimates provided are for the posterior mode (penalized maximum likelihood) for model parameters.
}
\references{
Kling, Baer, & Ackerly (2024). A tree's view of the terrain. Ecography e07131, https://doi.org/10.1111/ecog.07131
}
