% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{macroclimate}
\alias{macroclimate}
\title{Format macroclimate data}
\usage{
macroclimate(dem, interpolation = "bilinear")
}
\arguments{
\item{dem}{RasterLayer representing elevation, in meters.}

\item{interpolation}{Interpolation method. Either "bilinear" (default) or "ngb".}
}
\value{
A RasterBrick of macroclimate variables
}
\description{
Prepare a set of macroclimate rasters (biovariables bio12, bio1, bio5, and bio6), resampled and cropped to match resolution of elevation data.
}
