% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{moonshine}
\alias{moonshine}
\title{Example elevation raster}
\format{
\subsection{\code{moonshine}}{

A 10 m resolution RasterLayer with elevation values in meters.
}
}
\source{
Data from USGS at \url{https://apps.nationalmap.gov/downloader/#/}; documentation at \url{https://www.sciencebase.gov/catalog/item/5f7784fc82ce1d74e7d6cbec}
}
\usage{
moonshine
}
\description{
This raster DEM represents elevation variation over a rugged landscape near Moonshine Peak in southern Idaho, USA.
}
\keyword{datasets}
