% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mtpi}
\alias{mtpi}
\title{Multisacle TPI}
\usage{
mtpi(dem)
}
\arguments{
\item{dem}{RasterLayer representing elevation, in meters.}
}
\value{
A raster layer of mTPI
}
\description{
This function compuates multiscale topographic position index, sensu Theobald et al. (2015).
}
\references{
DM Theobald, D Harrison-Atlas, WB Monahan, CM Albano, Ecologically-relevant maps
of landforms and physiographic diversity for climate adaptation planning. PloS One e0143619 (2015)
}
