% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{northeast}
\alias{northeast}
\title{Northness and eastness}
\usage{
northeast(terr)
}
\arguments{
\item{terr}{RasterBrick with layers named "slope" and "aspect", e.g. generated with raster::terrain() function.}
}
\value{
A RasterBrick of with northness and eastness variables.
}
\description{
Calculate northness and eastness based on slope and aspect rasters.
}
