% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{windex}
\alias{windex}
\title{Windward exposure}
\usage{
windex(ne, terr)
}
\arguments{
\item{ne}{RasterBrick with layers named "northness" and "eastness", e.g. generated with northeast() function.}

\item{terr}{Raster object with a layer named "slope", e.g. generated with raster::terrain() function.}
}
\value{
A RasterLayer of windward exposure
}
\description{
Calculate windward exposure based on terrain rasters. Wind data are sourced from Kling & Ackerly (2020).
}
\references{
MM Kling, DD Ackerly, Global wind patterns and the vulnerability of wind-dispersed species to climate change. Nature Climate Change 10, 868–875 (2020).
}
